test(review-graph): enforce timer retention guard
Port the forced-GC graph replacement benchmark into Vitest and run it in CI.

Catch a single leaked 2,000-node graph with a 6 MiB bound.

refs #2073
