+++ .changelog/perf-2073-heap-retention.md
@@ -2,4 +2,4 @@
 section: Fixed
 ---
 
-- **Prove idle-eviction replacements release graph memory (refs #2073)** — Add a forced-GC benchmark for twenty 2,000-file-sized workspace graph replacements. The cache retains under 10 MiB after replacement, proving outgoing eviction timers do not retain prior graphs.
+- **Prove idle-eviction replacements release graph memory (closes #2073)** — Add an enforced forced-GC test for twenty 2,000-node workspace graph replacements. The cache retains under 6 MiB after replacement, proving outgoing eviction timers do not retain prior graphs.
